+++ apps/oidnDenoise.cpp
@@ -21,7 +21,7 @@ using namespace oidn;
 
 void printUsage()
 {
-  std::cout << "Intel(R) Open Image Denoise - Example" << std::endl;
+  std::cout << "Intel(R) Open Image Denoise" << std::endl;
   std::cout << "usage: oidnDenoise [-f/--filter RT|RTLightmap]" << std::endl
             << "                   [--ldr color.pfm] [--srgb] [--hdr color.pfm]" << std::endl
             << "                   [--alb albedo.pfm] [--nrm normal.pfm]" << std::endl
@@ -256,6 +256,9 @@ int main(int argc, char* argv[])
     }
 
     filter.commit();
+
+    if (verbose > 1 && inplace)
+      std::cout << "In-place mode" << std::endl;
 
     const double initTime = timer.query();