+++ cmake/oidn_dnnl.cmake
@@ -49,14 +49,6 @@ file(GLOB_RECURSE DNNL_SOURCES
   mkl-dnn/src/cpu/jit_avx2_conv_kernel_f32.cpp
   mkl-dnn/src/cpu/jit_avx2_convolution.hpp
   mkl-dnn/src/cpu/jit_avx2_convolution.cpp
-  mkl-dnn/src/cpu/jit_avx512_common_conv_kernel.hpp
-  mkl-dnn/src/cpu/jit_avx512_common_conv_kernel.cpp
-  mkl-dnn/src/cpu/jit_avx512_common_convolution.hpp
-  mkl-dnn/src/cpu/jit_avx512_common_convolution.cpp
-  mkl-dnn/src/cpu/jit_avx512_common_convolution_winograd.hpp
-  mkl-dnn/src/cpu/jit_avx512_common_convolution_winograd.cpp
-  mkl-dnn/src/cpu/jit_avx512_common_conv_winograd_kernel_f32.hpp
-  mkl-dnn/src/cpu/jit_avx512_common_conv_winograd_kernel_f32.cpp
   mkl-dnn/src/cpu/jit_avx512_core_f32_wino_conv_*.hpp
   mkl-dnn/src/cpu/jit_avx512_core_f32_wino_conv_*.cpp
   mkl-dnn/src/cpu/jit_generator.hpp
@@ -118,7 +110,7 @@ target_compile_definitions(dnnl
     -DDNNL_ENABLE_CONCURRENT_EXEC
 )
 
-set(DNNL_COMPILE_OPTIONS ${OIDN_ISA_FLAGS_SSE41})
+set(DNNL_COMPILE_OPTIONS "")
 if(WIN32 AND CMAKE_CXX_COMPILER_ID STREQUAL "Intel")
   # Correct 'jnl' macro/jit issue
   list(APPEND DNNL_COMPILE_OPTIONS "/Qlong-double")